fix: move BatchClient to utils
