--- src/batch/index.ts
@@ -1,2 +1,0 @@
-export * from './interface';
-export * from './default';
